feat: business 리팩토링+JWT

+++ business/views.py
@@ -58,19 +58,16 @@ def extract_chat_title(path: str) -> str:
 
 
 # Create your views here.
-class BusChatUploadView(APIView):
+class BusChatView(APIView):
     parser_classes = [MultiPartParser, FormParser]
-
     @swagger_auto_schema(
         operation_description="채팅 파일 업로드",
         manual_parameters=[
             openapi.Parameter(
-                "user_id",
-                openapi.IN_FORM,
-                type=openapi.TYPE_INTEGER,
-                required=True,
-                description="유저 ID",
-            ),
+                "Authorization",
+                openapi.IN_HEADER, 
+                description="access token", 
+                type=openapi.TYPE_STRING),
             openapi.Parameter(
                 "file",
                 openapi.IN_FORM,
@@ -79,19 +76,26 @@ class BusChatUploadView(APIView):
                 description="업로드할 채팅 파일",
             ),
         ],
+        responses={201: UploadResponseSerializer, 400: "Bad Request", 401: "Unauthorized"},
     )
     def post(self, request):
         serializer = ChatUploadRequestSerializer(data=request.data)
         if serializer.is_valid():
-            user_id = serializer.validated_data["user_id"]
             file = serializer.validated_data["file"]
+            author = request.user
+            
+            if not author.is_authenticated:
+                return Response(
+                    {"error": "User not authenticated"},
+                    status=status.HTTP_401_UNAUTHORIZED,
+                )
 
             # DB에 먼저 저장해서 경로를 얻는다
             chat = Chat.objects.create(
                 title="임시 제목",
                 content=file,
-                people_num=12,  # 필요시 동적으로 계산
-                user_id=User.objects.get(id=user_id),
+                people_num=12,  # 임시 값
+                user=request.user,
             )
 
             # 파일 경로에서 제목 추출
@@ -99,53 +103,78 @@ class BusChatUploadView(APIView):
             chat.title = extract_chat_title(file_path)
             chat.save()
 
-            response_serializer = UploadResponseSerializer(
-                {"chat_id_bus_contrib": chat.chat_id_bus_contrib}
+            response = UploadResponseSerializer(
+                {"chat_id": chat.chat_id}
             )
-            return Response(response_serializer.data, status=status.HTTP_201_CREATED)
+            return Response(response.data, status=status.HTTP_201_CREATED)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-class BusChatListView(APIView):
+            return Response(status=status.HTTP_400_BAD_REQUEST)
 
     @swagger_auto_schema(
         operation_id="채팅 목록 조회",
-        operation_description="채팅 목록을 조회합니다.",
-        responses={200: ListResponseSerializer, 404: "Not Found", 400: "Bad Request"},
+        operation_description="로그인된 유저의 채팅 목록을 조회합니다.",
+        manual_parameters=[
+            openapi.Parameter(
+                "Authorization",
+                openapi.IN_HEADER, 
+                description="access token", 
+                type=openapi.TYPE_STRING),
+        ],
+        responses={200: ListResponseSerializer(many=True), 404: "Not Found", 401: "Unauthorized"},
     )
-    def get(self, request, user_id):
-        try:
-            user = User.objects.get(id=user_id)
-            chats = Chat.objects.filter(user_id=user)
-
-            # Serialize the chat data
-            chat_data = [
-                {
-                    "id": chat.chat_id_bus_contrib,
-                    "title": chat.title,
-                    "people_num": chat.people_num,
-                    "uploaded_at": chat.updated_at,
-                }
-                for chat in chats
-            ]
-            return Response(chat_data, status=status.HTTP_200_OK)
-        except User.DoesNotExist:
+    def get(self, request):
+        author = request.user
+        if not author.is_authenticated:
             return Response(
-                {"error": "User not found"}, status=status.HTTP_404_NOT_FOUND
+                {"error": "User not authenticated"},
+                status=status.HTTP_401_UNAUTHORIZED,
             )
+
+        chats = Chat.objects.filter(user=author)
+        if not chats:
+            return Response(
+                {"error": "No chats found for this user"},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+        
+        # Serialize the chat data
+        chat_data = [
+            {
+                "chat_id": chat.chat_id,
+                "title": chat.title,
+                "people_num": chat.people_num,
+                "uploaded_at": chat.updated_at,
+            }
+            for chat in chats
+        ]
+        return Response(chat_data, status=status.HTTP_200_OK)
+
 
 
 class BusChatDetailView(APIView):
-
     @swagger_auto_schema(
-        operation_id="채팅 목록 조회",
-        operation_description="채팅 목록을 조회합니다.",
-        responses={200: ListResponseSerializer, 404: "Not Found", 400: "Bad Request"},
+        operation_id="채팅 삭제",
+        operation_description="채팅을 삭제합니다.",
+        manual_parameters=[
+            openapi.Parameter(
+                "Authorization",
+                openapi.IN_HEADER, 
+                description="access token", 
+                type=openapi.TYPE_STRING),
+        ],
+        responses={204: "No Content", 404: "Not Found", 400: "Bad Request"},
     )
     def delete(self, request, chat_id):
+        # authenticated user check
+        author = request.user
+        if not author.is_authenticated:
+            return Response(
+                {"error": "User not authenticated"},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
+        
         try:
-            chat = Chat.objects.get(chat_id_bus_contrib=chat_id)
+            chat = Chat.objects.get(chat_id=chat_id, user=author)
             chat.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         except Chat.DoesNotExist:
@@ -155,37 +184,54 @@ class BusChatDetailView(APIView):
 
 
 class BusChatAnalyzeView(APIView):
-    """
-    View to analyze chat data.
-    """
-
     @swagger_auto_schema(
         operation_id="채팅 분석",
         operation_description="채팅 데이터를 분석합니다.",
         request_body=ChatAnalysisRequestSerializer,
+        manual_parameters=[
+            openapi.Parameter(
+                "Authorization",
+                openapi.IN_HEADER, 
+                description="access token", 
+                type=openapi.TYPE_STRING),
+        ],
         responses={
             200: AnalyseResponseSerializer,
             404: "Not Found",
             400: "Bad Request",
+            403: "Forbidden"  # If the user does not have permission to analyze the chat
         },
     )
     def post(self, request, chat_id):
-        people_num = request.data.get("people_num")
-        rel = request.data.get("rel")
-        situation = request.data.get("situation")
-        analysis_start = request.data.get("analysis_start")
-        analysis_end = request.data.get("analysis_end")
-
-        if not all([people_num, rel, situation, analysis_start, analysis_end]):
+        # authenticated user check
+        author = request.user
+        if not author.is_authenticated:
             return Response(
-                {
-                    "detail": "[people_num, rel, situation, analysis_start, analysis_end] fields are required."
-                },
+                {"error": "User not authenticated"},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
+        
+        # Validate request data
+        serializer = ChatAnalysisRequestSerializer(data=request.data)
+        if serializer.is_valid() is False:
+            return Response(
+                {"error": "Invalid request data"},
                 status=status.HTTP_400_BAD_REQUEST,
             )
+        else:
+            people_num = serializer.validated_data["people_num"]
+            rel = serializer.validated_data["rel"]
+            situation = serializer.validated_data["situation"]
+            analysis_start = serializer.validated_data["analysis_start"]
+            analysis_end = serializer.validated_data["analysis_end"]
 
         try:
-            chat = Chat.objects.get(chat_id_bus_contrib=chat_id)
+            chat = Chat.objects.get(chat_id=chat_id)
+            if chat.user != author:
+                return Response(
+                    {"error": "You do not have permission to analyze this chat"},
+                    status=status.HTTP_403_FORBIDDEN,
+                )
         except Chat.DoesNotExist:
             return Response(
                 {"error": "Chat not found"}, status=status.HTTP_404_NOT_FOUND
@@ -203,32 +249,45 @@ class BusChatAnalyzeView(APIView):
             is_saved=1,
             analysis_date=timezone.now().date(),
             analysis_type="개인별 기여도 분석",
-            chat_id_bus_contrib=chat,
+            chat=chat,
         )
 
         return Response(
             {
-                "result_id_bus_contrib": result.result_id_bus_contrib,
+                "result_id": result.result_id,
             },
             status=status.HTTP_201_CREATED,
         )
 
 
 class BusResultListView(APIView):
-
     @swagger_auto_schema(
-        operation_id="채팅 분석",
-        operation_description="채팅 데이터를 분석합니다.",
+        operation_id="채팅 분석 결과 리스트 조회",
+        operation_description="로그인된 유저의 채팅 분석 결과 리스트를 조회합니다.",
+        manual_parameters=[
+            openapi.Parameter(
+                "Authorization",
+                openapi.IN_HEADER, 
+                description="access token", 
+                type=openapi.TYPE_STRING),
+        ],
         responses={200: AllResultSerializer, 404: "Not Found", 400: "Bad Request"},
     )
-    def get(self, request, user_id):
+    def get(self, request):
+        # authenticated user check
+        author = request.user
+        if not author.is_authenticated:
+            return Response(
+                {"error": "User not authenticated"},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
+        
+        # Get all analysis results for the logged-in user
         try:
-            user = User.objects.get(id=user_id)
-            results = ResultBusContrib.objects.filter(chat_id_bus_contrib__user_id=user)
-
+            results = ResultBusContrib.objects.filter(chat__user = author)
             result_data = [
                 {
-                    "result_id_bus_contrib": result.result_id_bus_contrib,
+                    "result_id": result.result_id,
                     "analysis_date": result.analysis_date,
                     "content": result.content,
                     "analysis_type": result.analysis_type,
@@ -237,9 +296,10 @@ class BusResultListView(APIView):
                 for result in results
             ]
             return Response(result_data, status=status.HTTP_200_OK)
-        except User.DoesNotExist:
+        except ResultBusContrib.DoesNotExist:
             return Response(
-                {"error": "User not found"}, status=status.HTTP_404_NOT_FOUND
+                {"error": "No analysis results found for this user"},
+                status=status.HTTP_404_NOT_FOUND,
             )
 
 
@@ -248,12 +308,30 @@ class BusResultDetailView(APIView):
     @swagger_auto_schema(
         operation_id="분석 결과 조회",
         operation_description="특정 분석 결과를 조회합니다.",
-        responses={200: DetailResultSerializer, 404: "Not Found", 400: "Bad Request"},
+        manual_parameters=[
+            openapi.Parameter(
+                "Authorization",
+                openapi.IN_HEADER, 
+                description="access token", 
+                type=openapi.TYPE_STRING),
+        ],
+        responses={200: DetailResultSerializer, 404: "Not Found", 400: "Bad Request", 401: "Unauthorized"},
     )
     def get(self, request, result_id):
+        # authenticated user check
+        author = request.user
+        if not author.is_authenticated:
+            return Response(
+                {"error": "User not authenticated"},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         try:
-            result = ResultBusContrib.objects.get(result_id_bus_contrib=result_id)
-
+            result = ResultBusContrib.objects.get(result_id=result_id)
+            if result.chat.user != author:
+                return Response(
+                    {"error": "You do not have permission to view this analysis result"},
+                    status=status.HTTP_403_FORBIDDEN,
+                )
             return Response({"content": result.content}, status=status.HTTP_200_OK)
         except ResultBusContrib.DoesNotExist:
             return Response(
@@ -263,11 +341,30 @@ class BusResultDetailView(APIView):
     @swagger_auto_schema(
         operation_id="분석 결과 삭제",
         operation_description="특정 분석 결과를 삭제합니다.",
-        responses={204: "No Content", 404: "Not Found"},
+        manual_parameters=[
+            openapi.Parameter(
+                "Authorization",
+                openapi.IN_HEADER, 
+                description="access token", 
+                type=openapi.TYPE_STRING),
+        ],
+        responses={204: "No Content", 404: "Not Found", 401: "Unauthorized", 403: "Forbidden"},
     )
     def delete(self, request, result_id):
+        # authenticated user check
+        author = request.user
+        if not author.is_authenticated:
+            return Response(
+                {"error": "User not authenticated"},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         try:
-            result = ResultBusContrib.objects.get(result_id_bus_contrib=result_id)
+            result = ResultBusContrib.objects.get(result_id=result_id)
+            if result.chat.user != author:
+                return Response(
+                    {"error": "You do not have permission to delete this analysis result"},
+                    status=status.HTTP_403_FORBIDDEN,
+                )
             result.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         except ResultBusContrib.DoesNotExist:
